+++ scm-webapp/src/main/java/sonia/scm/api/v2/resources/BranchCollectionToDtoMapper.java
@@ -30,7 +30,6 @@ import de.otto.edison.hal.HalRepresentation;
 import de.otto.edison.hal.Link;
 import de.otto.edison.hal.Links;
 import sonia.scm.repository.Branch;
-import sonia.scm.repository.NamespaceAndName;
 import sonia.scm.repository.Repository;
 import sonia.scm.repository.RepositoryPermissions;
 
@@ -55,11 +54,11 @@ public class BranchCollectionToDtoMapper {
   public HalRepresentation map(Repository repository, Collection<Branch> branches) {
     return new HalRepresentation(
       createLinks(repository),
-      embedDtos(getBranchDtoList(repository.getNamespace(), repository.getName(), branches)));
+      embedDtos(getBranchDtoList(repository, branches)));
   }
 
-  public List<BranchDto> getBranchDtoList(String namespace, String name, Collection<Branch> branches) {
-    return branches.stream().map(branch -> branchToDtoMapper.map(branch, new NamespaceAndName(namespace, name))).collect(toList());
+  public List<BranchDto> getBranchDtoList(Repository repository, Collection<Branch> branches) {
+    return branches.stream().map(branch -> branchToDtoMapper.map(branch, repository)).collect(toList());
   }
 
   private Links createLinks(Repository repository) {